backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
sqlparse==0.4.3
tzdata==2022.5